requirements.txt
asttokens==2.4.0
async-generator==1.10
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.3
bokeh==3.3.0
branca==0.6.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
colorclass==2.2.2
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.4
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
docopt==0.6.2
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.3
executing==2.0.0
ffpyplayer==4.5.1
Flask==3.0.0
folium==0.14.0
fonttools==4.43.1
gevent==23.9.1
gevent-websocket==0.10.1
gmplot==1.4.1
greenlet==3.0.0
h11==0.14.0
idna==3.4
imutils==0.5.4
ipykernel==6.25.2
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pip-upgrader==1.4.15
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
scipy==1.11.3
selenium==4.14.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
traitlets==5.11.2
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
websocket-client==1.6.4
Werkzeug==3.0.0
wsproto==1.2.0
xyzservices==2023.10.0
zope.event==5.0
zope.interface==6.1
